dom/src/threads/test/test_chromeWorker.html

<html>
<head>
  <title>Test for DOM Worker Threads</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

  try {
    var worker = new ChromeWorker("simpleThread_worker.js");
    ok(false, "ChromeWorker constructor should be blocked!");
  }
  catch (e) {
    ok(true, "ChromeWorker constructor wasn't blocked!");
  }

  try {
    var worker = Components.classes["@mozilla.org/threads/workerfactory;1"]
      .createInstance(Components.interfaces.nsIWorkerFactory)
      .newChromeWorker("simpleThread_worker.js");
    ok(false, "nsIWorkerFactory should be blocked!");
  }
  catch (e) {
    ok(true, "nsIWorkerFactory wasn't blocked!");
  }

</script>
</pre>
</body>
</html>
